Finite Set as Color Grid
First look at visualing a domain, codomain, and a mapping between them. The view folder was added to house this view into Finite Sets, as well as any others that come. The public folder is where static documents are stored, and the root of localhost when 'npm start' is executed.

Run 'npm run build' to compile ./index.js for the browser via browserify. Run 'npm run watch' to watch for changes via watchify.